backend/requirements.txt
asgiref==3.8.1
attrs==24.2.0
black==22.3.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.0.4
click-plugins==1.1.1
cligj==0.7.2
Django==5.0.6
django-colorfield==0.11.0
django-cors-headers==4.4.0
django-environ==0.11.2
django-imagefield==0.19.1
django-model-utils==4.5.1
django-smart-selects==1.6.0
djangorestframework==3.15.1
environ==1.0
fiona==1.9.6
geobr==0.2.1
geopandas==0.14.4
gunicorn==22.0.0
idna==3.8
lxml==5.3.0
mypy-extensions==1.0.0
numpy==2.1.0
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
psycopg2-binary==2.9.9
pyproj==3.6.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
Shapely==1.8.5.post1
six==1.16.0
sqlparse==0.5.0
swapper==1.4.0
tomli==2.0.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
